% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_table_one.R
\name{make_table_one}
\alias{make_table_one}
\title{Tableone wrapper based off of gtsummary}
\usage{
make_table_one(data, table_vars, by)
}
\arguments{
\item{by}{}
}
\value{

}
\description{
Tableone wrapper based off of gtsummary
}
\examples{
table_one(data = df, table_vars = table_vars, by = ln_positive)

}
