% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_date.R
\name{save_date}
\alias{save_date}
\title{Append today's date to the file that is being saved}
\usage{
save_date(file_name)
}
\arguments{
\item{file_name}{a character file name the date will be appended onto}
}
\value{

}
\description{
Append today's date to the file that is being saved
}
\examples{

save_date('new_plot.jpg')
}
