% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_analysis_folders.R
\name{setup_analysis_folders}
\alias{setup_analysis_folders}
\title{Setup data analysis folder structure}
\usage{
setup_analysis_folders(
  folders = c("src", "reports", "figures", "data", "docs", "munge", "tables", "output")
)
}
\value{

}
\description{
Creates standard src, reports, figures, data, munge, and docs folder
}
\examples{

}
