% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_plot.R
\name{surv_plot}
\alias{surv_plot}
\title{Create a Kaplan-Meier Plot with Corresponding 'at risk' table}
\usage{
surv_plot(
  data,
  fit,
  break_x_by = 12,
  xlim = NULL,
  plot_table_size = 5,
  expand_plot = 0,
  curve_labels = NULL,
  curve_text_size = 5
)
}
\arguments{
\item{data}{a tibble or dataframe}

\item{fit}{a survfit model}

\item{break_x_by}{sequence for the x axis}

\item{xlim}{limits for the x axis}

\item{plot_table_size}{size of the table}

\item{expand_plot}{expand the plot to provide more room}

\item{curve_labels}{labels for the legend of the curve}

\item{line_annotation_size}{size of the text for the annotation}
}
\value{

}
\description{
Create a Kaplan-Meier Plot with Corresponding 'at risk' table
}
\examples{

}
