% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_cox_model.R
\name{tidy_cox_model}
\alias{tidy_cox_model}
\title{Tidy cox model results}
\usage{
tidy_cox_model(fit, exponentiate = T, conf.int = T)
}
\arguments{
\item{fit}{cox fit object}

\item{exponentiate}{exponentiate the results}
}
\value{

}
\description{
Tidy cox model results
}
\examples{



}
