% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_plot2.R
\name{tidy_summary_survfit}
\alias{tidy_summary_survfit}
\title{Helper function to extract survfit summary objects}
\usage{
tidy_summary_survfit(fit)
}
\arguments{
\item{fit}{summary survfit object}
}
\value{

}
\description{
Helper function to extract survfit summary objects
}
\examples{

tidy_summary_survfit


}
