% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_surv_table.R
\name{tidy_surv_table}
\alias{tidy_surv_table}
\title{Tidy survival estimates into GT tables}
\usage{
tidy_surv_table(fit, times)
}
\arguments{
\item{fit}{survfit object}

\item{times}{sequences of times to generate the survival estimates for}
}
\value{

}
\description{
Tidy survival estimates into GT tables
}
