% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_plot.R
\name{surv_plot}
\alias{surv_plot}
\title{Title}
\usage{
surv_plot(
  data,
  fit,
  break_x_by = 12,
  xlim = NULL,
  curve_labels = NULL,
  curve_text_size = 5,
  logrank_test = T,
  table_title = "No. at risk",
  table_title_size = 15,
  table_text_size = 5,
  table_label_spacing = 20,
  plot_table_ratio = c(0.9, 0.1),
  pvalue_text_size = 5,
  pvalue_text_location = c(0.02, 0.02),
  curve_color_palette = ggsci::pal_d3(),
  expand_plot_area = 12,
  theme_plot_text_size = 15,
  plot_margin = margin(10, 50, 0, 50),
  table_margin = margin(0, 0, 15, 0),
  xlab = "Time, mo",
  ylab = "Overall Survival, \%"
)
}
\arguments{
\item{theme_plot_text_size}{}
}
\value{

}
\description{
Title
}
